Fix issue with react-native's use_native_modules!
Version 0.60 of react-native introduces a feature called autolinking. It
works by importing a bundled script into the project's Podfile and
running the function use_native_modules!. That function depends on the
Pod::UI module, which is not present when only importing cocoapods-core,
causing the project scanner to crash when parsing such podfiles.